% Generated by roxygen2 (4.0.2): do not edit by hand
\name{importPir}
\alias{importPir}
\title{Import Pirouette File.}
\usage{
importPir(dataFile)
}
\arguments{
\item{dataFile}{Character, a valid path to a Pirouette data file.}
}
\value{
A list consisting  of 5 elements.
}
\description{
Imports directly from a Pirouette file and gives back all the
		data in a list consisting of 5 elements.
}
\details{


Here the details on how this function is working, what it does etc.
}
\examples{
\dontrun{
	pirData <- importPir("~/Documents/pathToData.pir")
	}
}

