% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{readPir}
\alias{readPir}
\alias{readPir-package}
\title{Read in Pirouette Files}
\description{
Package 'readPir' allows you to directly read and import data from a Pirouette File.
}
\details{
This here are the details. Are there any details to mention?
}
\section{Author}{
 Santa Claus, Kermet the Frog
}

\section{Maintainer}{
 Santa Claus <santa.claus@gmail.com>
}

\section{References}{
 Here some link to a paper etc.
}
\seealso{
\code{\link{aquaphotomics::gdmm}}
}

